backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose
passlib
bcrypt
pydantic
pydantic-settings
python-multipart
python-dotenv
httpx
